// src/app/models/course-management/swuCurriculumData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SwuCurriculumData {
    /** ชื่อหลักสูตร (ภาษาอังกฤษ) */
    curriculumNameEn?: string;
    /** ชื่อหลักสูตร (ภาษาไทย) */
    curriculumNameTh?: string;
    /** PK รหัสหลักสูตร มศว Not Autorun  */
    curriculumSwuId?: string;
    first?: number;
    mode?: string;
    /** หน่วยงานเจ้าของหลักสูตร (อังกฤษ) */
    ownerDepNameEn?: string;
    /** หน่วยงานเจ้าของหลักสูตร (ไทย) */
    ownerDepNameTh?: string;
    rowNum?: number;
    /** รูปแบบการค้นหารายวิชาเทียบเคียง */
    searchFormat?: number;
    size?: number;
    /** custom รหัสวิชา (ภาษาอังกฤษ) */
    subjectCodeEn?: string;
    /** custom รหัสวิชา (ภาษาไทย) */
    subjectCodeTh?: string;
    /** custom จํานวนหน่วยกิต */
    subjectCredit?: number;
    /** Unique Key */
    subjectKey?: string;
    /** custom ชื่อรายวิชา (ภาษาอังกฤษ) */
    subjectNameEn?: string;
    /** custom ชื่อรายวิชา (ภาษาไทย) */
    subjectNameTh?: string;
    /** custom ชุดวิชา */
    subjectSet?: string;
    /** custom FK (swu_subject) */
    subjectSwuId?: number;
    curriculumInput?:string;
    courseInput?:string;
    courseNameTh?:string;
    courseNameEn?:string;
    creditAmount?:number;

    subjectCreditFormat?: string;

}